Resolve compiler warnings

+++ include/kheap.h
@@ -3,10 +3,10 @@
 
 #include <stdint.h>
 
-static uint32_t _kmalloc_int(uint32_t size, uint32_t * phys, uint8_t align);
-uint32_t kmalloc_a(uint32_t size);
-uint32_t kmalloc_p(uint32_t size, uint32_t * phys);
-uint32_t kmalloc_ap(uint32_t size, uint32_t * phys);
-uint32_t kmalloc(uint32_t size);
+static uintptr_t _kmalloc_int(uint32_t size, uint32_t * phys, uint8_t align);
+uintptr_t kmalloc_a(uint32_t size);
+uintptr_t kmalloc_p(uint32_t size, uint32_t * phys);
+uintptr_t kmalloc_ap(uint32_t size, uint32_t * phys);
+uintptr_t kmalloc(uint32_t size);
 
 #endif
